Change to a different DS3231 lib
While Adafruit's RTClib is easy to work with, there's a lot that
goes on under the hood, which causes problems when going in and
out of deep sleep. The switch to a less in depth library should
make it easier to manage the DS3231 in and out of deep sleep.