Change from RTClib to another DS3231 RTC library
This change is an attempt to save some memory by using a smaller
library, though the more complicated code may not be worth it.
The startTime functionality is broken in this version of the code,
but everything else works as before, with the added limitation
that sleep / sample durations (if set) cannot be longer than 1
day. This is because I would have to implement the datetime
calculations myself.