Integrate ULP into program to read from MS5803
This update makes use of the ULP and a bit banged implementation
of I2C to read from the MS5803 and buffer data. This reduces how
frequently the ESP32's main processor needs to be active and
should offer a significant improvement to the battery life.

In order to make use of this code, the hardware must be updated
to connect GPIO 0 to SCL and GPIO 14 to SDA of the ESP32's
hardware I2C bus. These new pins are used as the bit banged SCL
and SDA lines respectively, and they need to be attached to the
existing devices.