Implement a testing version of ULP time syncing
In its current state the program only runs the ULP once and monitors the
time between samples being detected. The ULP program has been modifed to
compensate for time drift using the DS3231 alarm.

The device initialization code has also been slightly modified to clear
the DS3231 alarm before starting the ULP. It now also waits before
starting the ULP program according to the sampling rate (e.g. at 4Hz we
make sure we wait 250ms before starting the ULP). This is done to
prevent time sync issues: if we started it right away, the ULP would
wake up around 750ms for the 4th sample and need to wait 250ms for the
alarm, causing it to think it is way out of sync. This is only an issue
when first starting the ULP, since it sleeps between samples, just not
when starting the program.