Only write last temperature sample for each data dump (#26)
Previously, a corresponding temperature reading would be written with
each pressure reading. However, since the pressure is already
temperature compensated, it hasn't proved to be a useful metric to track
at the same frequency as pressure and wastes an extra byte of storage
with each sample.
Now, a single temperature reading will be written to file corresponding
to the last pressure sample in the buffer (i.e. the last sample before
the main core wakes up to dump data to the SD card)