Remove timezone stuff for now (#39)
While it was nice to have accurate time information on files written to
the SD card, I don't want to deal with passing this info from the Web
dashboard at this time. I would need a bit more of a robust setup to
make it nicer to implement. It also might overcomplicate outputted data.