Implement ULP sleep time adjustments and fix DS3231 alarm time syncing bug
Per #32, the ULP's sleep duration is now adjusted based on the most
recent number of ULP wait loop iterations each time the main core wakes
up. If a substantial adjustment is needed (the adjustement is more than
half the time the ULP is expected to be in the wait loop for a given
sample), then it is reduced to prevent a sample from being missed.

A major bug was discovered with the ULP seeming to fall out of sync each
time it would enter deep sleep after the first time. It seems this was
caused by failing to run hulp_peripherals_on() in the deep sleep wakeup
code in order to properly keep RTC peripherals on. This function is run
in ulp_init(), but it seems to "forget" this configuration when it wakes
up after deep sleep.

GPIO functions were switched from Arduino style (pinMode, digitalWrite,
etc.) to ESP IDF style (specifically rtc_gpio_* functions) to avoid
needing to run GPIO configuration each time the device wakes up (now it
only runs once when the device is powered on).

Other bug fixes:
- In the ULP program: when adjusting the number of wait loop
  iterations, we were right shifting by samples_per_second instead of
  log2(samples_per_second). This is now fixed.
- The number of ULP processor cycles required to execute a single
  iteration of either wait loop in the ULP program was incorrect. This
  was caused by failing to account for the faster instruction fetch time
  of instructions following ALU and branch instructions. Instead of the
  usual 4 cycles, it is only 2 cycles to fetch in this case. This was
  confirmed by timing these loops with an oscilloscope.